Update order input fields to accept decimal and fractional measurements
Modify width and height input fields in the new order form to accept text input and parse it as decimal or fractional numbers using a helper function. Update the `orders` table schema and database migration to store width and height as `numeric` instead of `integer`.

+++ shared/schema.ts
@@ -22,8 +22,8 @@ export const orders = pgTable("orders", {
   // Frame Details
   frameSku: text("frame_sku").notNull(),
   chopOnly: boolean("chop_only").notNull().default(false),
-  width: integer("width").notNull(),
-  height: integer("height").notNull(),
+  width: decimal("width").notNull(),
+  height: decimal("height").notNull(),
   
   // Mat Configuration
   matBorderAll: decimal("mat_border_all"),
